Extract Within to the Cycles namespace
