+++ lib/sof/cycle.rb
@@ -204,25 +204,6 @@ module SOF
 
     def as_json(...) = notation
 
-    class Within < self
-      @volume_only = false
-      @notation_id = "W"
-      @kind = :within
-      @valid_periods = %w[D W M Y]
-
-      def to_s = "#{volume}x within #{date_range}"
-
-      def date_range
-        return humanized_span unless active?
-
-        [start_date, final_date].map { _1.to_fs(:american) }.join(" - ")
-      end
-
-      def final_date(_ = nil) = time_span.end_date(start_date)
-
-      def start_date(_ = nil) = from_date.to_date
-    end
-
     class VolumeOnly < self
       @volume_only = true
       @notation_id = nil
